+++ functions/mi_hub_call_history_functions.py
@@ -1,5 +1,6 @@
 from data_sources.blaise_api import get_list_of_installed_instruments, get_instrument_data
-from functions.call_history_functions import get_cati_mi_hub_call_history, merge_cati_call_history_and_instrument_data
+from data_sources.cati_database import get_cati_mi_hub_call_history_from_database
+from functions.call_history_functions import merge_cati_call_history_and_instrument_data, get_instrument_name_from_id
 from functions.csv_functions import write_list_of_dicts_to_csv
 from functions.folder_functions import create_instrument_name_folder_in_tmp_directory, get_tmp_directory_path
 from models.mi_hub_call_history import MiHubCallHistory
@@ -25,6 +26,29 @@ def get_mi_hub_call_history(config):
         create_instrument_name_folder_in_tmp_directory(instrument_name)
         csv_file = f"{tmp_folder}/{instrument_name}/call_history.csv"
         write_list_of_dicts_to_csv(csv_file, mi_hub_call_history_grouped[instrument_name], MiHubCallHistory.fields())
+
+
+def get_cati_mi_hub_call_history(config, instrument_list):
+    results = get_cati_mi_hub_call_history_from_database(config)
+    cati_mi_hub_call_history_list = []
+    for item in results:
+        cati_mi_hub_call_history = MiHubCallHistory(
+            questionnaire_id=item.get("InstrumentId"),
+            serial_number=item.get("PrimaryKeyValue"),
+            call_number=item.get("CallNumber"),
+            dial_number=item.get("DialNumber"),
+            interviewer=item.get("Interviewer"),
+            dial_result=item.get("DialResult"),
+            dial_line_number=item.get("DialedNumber"),
+            seconds_interview=item.get("dial_secs")
+        )
+        cati_mi_hub_call_history.generate_dial_date_and_time_fields(item.get("StartTime"), item.get("EndTime"))
+        instrument_name = get_instrument_name_from_id(cati_mi_hub_call_history.questionnaire_id, instrument_list)
+        if instrument_name != "":
+            cati_mi_hub_call_history.questionnaire_name = instrument_name
+        cati_mi_hub_call_history_list.append(cati_mi_hub_call_history)
+    print(f"Found {len(results)} mi hub call history records in the CATI database")
+    return cati_mi_hub_call_history_list
 
 
 def group_by_instrument(data):
